Delete splits/external pre-processing files/comments
