+++ recipes/Fisher-Callhome-Spanish/ST/transformer/hparams/conformer.yaml
@@ -60,7 +60,7 @@ speed_perturb: !new:speechbrain.processing.speech_augmentation.SpeedPerturb
     speeds: [90, 100, 110]
 
 # Trainer settings
-number_of_epochs: 30
+number_of_epochs: 40
 valid_search_eopch: 10
 batch_size: 4 # this works for 2 GPUs with 11GB
 gradient_accumulation: 16
