--- recipes/Fisher-Callhome-Spanish/Tokenizer/pre_processing.py
@@ -1,1 +1,0 @@
-../pre_processing.py
