+++ recipes/Fisher-Callhome-Spanish/fisher_callhome_prepare.py
@@ -7,7 +7,9 @@ YAO-FEI, CHENG 2021
 """
 
 import os
+import re
 import json
+import string
 import logging
 import subprocess
 
@@ -16,13 +18,6 @@ from dataclasses import dataclass, field
 
 import torch
 import torchaudio
-
-from pre_processing import (
-    clean_transcription,
-    clean_translation,
-    normalize_punctuation,
-    remove_punctuation,
-)
 
 from tqdm import tqdm
 from speechbrain.utils.data_utils import get_all_files
@@ -120,6 +115,8 @@ def prepare_fisher_callhome_spanish(
 
     corpus_path = f"{save_folder}/fisher-callhome-corpus"
     download_translations(path=corpus_path)
+
+    make_data_splits()  # make splitted data list from mapping files
 
     for dataset in datasets:
         if not os.path.exists(f"{save_folder}/{dataset}/wav"):
@@ -459,6 +456,273 @@ def download_translations(path: str):
         logger.info(f"Translation file not found. Downloading from {repo}.")
         subprocess.run(["git", "clone", repo])
         subprocess.run(["mv", "fisher-callhome-corpus", f"{path}"])
+
+
+def make_data_splits(
+    mapping_folder: str = "./data/fisher-callhome-corpus/mapping",
+):
+    """make data split from mapping file"""
+    fisher_splits = ["dev", "dev2", "test", "train"]
+
+    if not os.path.exists("splits"):
+        os.mkdir("splits")
+
+    for fisher_split in fisher_splits:
+        split = set()
+        with open(
+            f"{mapping_folder}/fisher_{fisher_split}", "r", encoding="utf-8"
+        ) as fisher_file, open(
+            f"./splits/{fisher_split}", "a+", encoding="utf-8"
+        ) as split_file:
+            fisher_file_lines = fisher_file.readlines()
+
+            for fisher_file_line in fisher_file_lines:
+                fisher_file_line = fisher_file_line.strip()
+                fisher_file_id = fisher_file_line.split(" ")[0]
+                split.add(fisher_file_id)
+
+            split = sorted(list(split))
+            for file_id in split:
+                split_file.write(f"{file_id}\n")
+
+
+def remove_punctuation(text: str) -> str:
+    """remove punctuation from given string"""
+
+    # remove punctuation except apostrophe
+    text = text.replace("<space>", "spacemark")
+    text = text.replace("'", "apostrophe")
+
+    # based on the definition of [[:punct]]
+    punctuation = r"[{}]".format(string.punctuation)
+
+    text = re.sub(punctuation, "", text)
+    text = text.replace("spacemark", "<space>")
+    text = text.replace("apostrophe", "'")
+
+    # remove consecutive commas and spaces
+    text = re.sub(r"\s+", " ", text)
+    text = re.sub(f"^\\s+", "", text)
+    text = re.sub(r"\s+$", "", text)
+
+    return text
+
+
+def normalize_punctuation(text: str) -> str:
+    """remove punctuation from given string"""
+
+    # remove brachets and inside
+    text = re.sub(r"\([^)]*\)", " ", text)
+    text = re.sub(r"\[[^]]+\]", " ", text)
+
+    # normalize punctuation
+    text = re.sub(r"_", "", text)
+    text = re.sub(r"`", "'", text)  # for En
+    text = re.sub(r"´", "'", text)  # for En
+    text = re.sub(r"\¨", "'", text)  # I¨m -> I'm etc.
+
+    # remove noisy parts
+    text = re.sub(r"noise", "", text)
+    text = re.sub(r"laughter", "", text)
+    text = re.sub(r"background noise", "", text)
+    text = re.sub(r"background speech", "", text)
+
+    # fisher_train
+    text = re.sub(r"i\/he", "i", text)
+    text = re.sub(r"i\/she", "i", text)
+    text = re.sub(r" \/\?", "\\?", text)
+    text = re.sub(r" \/ ", " ", text)
+    text = re.sub(r"a\/c", "", text)
+    text = re.sub(r"stay\/", "stay", text)
+    text = re.sub(r"boys\/", "boys", text)
+    text = re.sub(r"right\/", "right", text)
+    text = re.sub(r"follow\/", "follow", text)
+    text = re.sub(r"Jose\/Josefina", "Jose", text)
+    text = re.sub(r"welfare\/foreign", "welfare", text)
+    text = re.sub(r"\<foreign lang=\"English\"", "", text)
+    text = re.sub(r"\/foreign/", "", text)
+    text = re.sub(r"\<plural\>", "", text)
+    text = re.sub(r"\<barely makes any sense\>", "", text)
+    text = re.sub(r"\<kind of a weird phrase\>", "", text)
+    text = re.sub(r"\<last word does not fit there\>", "", text)
+    text = re.sub(r"\<players with the meaning of singers\>", "", text)
+    text = re.sub(r"\<this phrase barely made any sense whatsoever\>", "", text)
+    text = re.sub(
+        r"\<colorcito does not exist as a word so I have no ideea what he means about that\>",
+        "",
+        text,
+    )
+    text = re.sub(r"\<foreign", "", text)
+    text = re.sub(r"foreign\>", "", text)
+
+    # fisher_dev
+    text = re.sub(r"her\/his", "her", text)
+    text = re.sub(r"o\/", "o", text)
+    text = re.sub(r"co\/", "co", text)
+    text = re.sub(r"L \/ ", "", text)
+    text = re.sub(r"\<\?\?\?\>", "", text)
+    text = re.sub(r"\<from Texas\>", "", text)
+    text = re.sub(r"\<weird phrase\>", "", text)
+    text = re.sub(r"\<weird phrase\>", "", text)
+    text = re.sub(r"\<this makes no sense\>", "", text)
+    text = re.sub(r"Salvador\>", "Salvador", text)
+
+    # fisher_dev 2
+    text = re.sub(r"A\/C", "", text)
+    text = re.sub(r"She\/he", "She", text)
+    text = re.sub(r"you\/he", "you", text)
+    text = re.sub(r"you\/she", "you", text)
+    text = re.sub(r"Um\/", "Um", text)
+    text = re.sub(r"name\/", "name", text)
+    text = re.sub(r"American\/", "American", text)
+    text = re.sub(r"\<\?\>", "", text)
+    text = re.sub(r"\<metaphoric meaning\>", "", text)
+    text = re.sub(r"\<missing text \? \>", "", text)
+    text = re.sub(
+        r"\<broken phrase but I tried to guess what would it mean if it was complete\>",
+        "",
+        text,
+    )
+
+    # fisher_test
+    text = re.sub(r"she\/he", "she", text)
+    text = re.sub(r"her\/him", "her", text)
+    text = re.sub(r"is\/", "is", text)
+    text = re.sub(r"and\/or", "and", text)
+    text = re.sub(r"Then\/Well", "Then", text)
+    text = re.sub(r"fine\/well", "fine", text)
+    text = re.sub(r"Likewise\/Equally", "Likewise", text)
+    text = re.sub(r"boyfriend\/girlfriend", "boyfriend", text)
+    text = re.sub(r"living room \/ dining room", "living room", text)
+    text = re.sub(r"\<very bad phrase\>", "", text)
+    text = re.sub(r"\<poorly written phrase\>", "", text)
+    text = re.sub(r"\<this phrase barely even made sense\>", "", text)
+    text = re.sub(
+        r"\<very poorly written phrase but I think this is what was supposed to mean\>",
+        "",
+        text,
+    )
+    text = re.sub(r"what\)\)", "what", text)
+
+    # remove noisy punctuation
+    text = re.sub(r"\(", " ", text)
+    text = re.sub(r"\)", " ", text)
+    text = re.sub(r"\<", " ", text)
+    text = re.sub(r"\>", " ", text)
+    text = re.sub(r"\[", " ", text)
+    text = re.sub(r"\]", " ", text)
+    text = re.sub(r"\{", " ", text)
+    text = re.sub(r"\}", " ", text)
+    text = re.sub(r"\\", " ", text)
+    text = re.sub(r"\/", " ", text)
+    text = re.sub(r"\;", " ", text)
+    text = re.sub(r"~", " ", text)
+    text = re.sub(r"=", " ", text)
+    text = re.sub(r"\·", " ", text)
+
+    text = re.sub(r"^\.\s*$", "", text)  # only period sentence
+    text = re.sub(r"^\?\s*$", "", text)  # only question mark sentence
+    text = re.sub(r"\s+", " ", text)  # remove consecutive whitespaces
+
+    # remove the first and last whitespaces
+    text = re.sub(r"^\s+", "", text)
+    text = re.sub(r"\s+$", "", text)
+
+    text = text.lstrip()
+
+    return text
+
+
+def clean_transcription(transcription: str) -> str:
+    """
+    original: https://github.com/jamfly/AlloST/blob/main/egs/fisher_callhome_spanish/st1/local/fsp_data_prep.sh
+
+    clean a given transcription and return a cleaned transcription
+    """
+
+    transcription = transcription.replace("</", "lendarrow")
+    transcription = transcription.replace("<", "larrow")
+    transcription = transcription.replace(">", "rarrow")
+
+    punctuation = r"[{}]".format(string.punctuation)
+    transcription = re.sub(punctuation, "", transcription)
+
+    transcription = transcription.replace("larrow", "<")
+    transcription = transcription.replace("rarrow", ">")
+    transcription = transcription.replace("lendarrow", "</")
+
+    transcription = transcription.replace("Á", "á")
+    transcription = transcription.replace("Í", "í")
+    transcription = transcription.replace("Ó", "ó")
+    transcription = transcription.replace("Ú", "ú")
+    transcription = transcription.replace("¨", "")
+    transcription = transcription.replace("·", "")
+    transcription = transcription.replace("´", "")
+    transcription = transcription.replace("¿", "")
+    transcription = transcription.replace("¡", "")
+    transcription = transcription.replace("N", "n")
+
+    transcription = transcription.lower()
+
+    transcription = remove_labels(transcription)
+
+    return transcription
+
+
+def clean_translation(translation: str) -> str:
+    """clean a given translation and returne a cleaned translation"""
+    translation = translation.strip()
+    translation = translation.lower()
+
+    translation = translation.replace("¿", "")
+    translation = translation.replace("¡", "")
+
+    return translation
+
+
+def remove_labels(transcription: str):
+    """remove label such as <laugh> from transcript"""
+
+    transcription = re.sub(
+        r"<\s*[/]*\s*\s*for[ei][ei]g[nh]\s*\w*>", "", transcription
+    )
+    transcriptions = re.findall(r"<lname>\([^<]*\)<\/lname>", transcription)
+
+    if len(transcriptions) > 0:
+        transcription = transcriptions[0]
+
+    transcription = re.sub(r"<lname[\/]*>", "", transcription)
+    transcription = re.sub(r"<laugh>", "", transcription)
+    transcription = re.sub(r"<\/laugh>", "", transcription)
+    transcription = re.sub(r"<\s*cough[\/]*>", "[noise]", transcription)
+    transcription = re.sub(r"<sneeze[\/]*>", "[noise]", transcription)
+    transcription = re.sub(r"<breath[\/]*>", "[noise]", transcription)
+    transcription = re.sub(r"<lipsmack[\/]*>", "[noise]", transcription)
+    transcription = re.sub(r"<background>", "", transcription)
+    transcription = re.sub(r"<\/background>", "", transcription)
+    transcription = re.sub(r"<[/]?background[/]?>", "[noise]", transcription)
+    transcription = re.sub(r"<laugh>", "", transcription)
+    transcription = re.sub(r"<\/laugh>", "", transcription)
+    transcription = re.sub(r"<[/]?laugh[/]?>", "[laughter]", transcription)
+    transcription = re.sub(r"<foreign langenglishhip hop", "", transcription)
+    transcription = re.sub(r"<foreign langenglishonline", "", transcription)
+    transcription = re.sub(r"<foreign langenglish", "", transcription)
+    transcription = re.sub(r"</foreign", "", transcription)
+    transcription = re.sub(r"<[/]?foreing\s*\w*>", "", transcription)
+    transcription = re.sub(r"</b", "", transcription)
+    transcription = re.sub(r"<foreign langengullís>", "", transcription)
+    transcription = re.sub(r"foreign>", "", transcription)
+    transcription = re.sub(r">", "", transcription)
+
+    is_match = re.search(r"\(\)", transcription)
+
+    if is_match is not True:
+        transcription = re.sub(r"\[noise\]", "", transcription)
+        transcription = re.sub(r"\[laughter\]", "", transcription)
+        transcription = re.sub(r"^\s\s*|\s\s*$", "", transcription)
+        transcription = re.sub(r"^\s\s*", " ", transcription)
+
+    return transcription
 
 
 if __name__ == "__main__":
