--- recipes/Fisher-Callhome-Spanish/splits/dev
@@ -1,20 +1,0 @@
-20051009_182032_217_fsp.sph
-20051009_210519_219_fsp.sph
-20051010_212418_225_fsp.sph
-20051016_180547_265_fsp.sph
-20051016_210626_267_fsp.sph
-20051017_180712_270_fsp.sph
-20051017_220530_275_fsp.sph
-20051017_234550_276_fsp.sph
-20051018_210220_279_fsp.sph
-20051018_210744_280_fsp.sph
-20051019_190221_288_fsp.sph
-20051019_210146_289_fsp.sph
-20051019_230329_292_fsp.sph
-20051022_180817_311_fsp.sph
-20051023_232057_325_fsp.sph
-20051024_180453_327_fsp.sph
-20051024_181110_329_fsp.sph
-20051025_212334_337_fsp.sph
-20051026_180724_341_fsp.sph
-20051026_211309_346_fsp.sph
