--- recipes/Fisher-Callhome-Spanish/splits/dev2
@@ -1,20 +1,0 @@
-20050909_210655_26_fsp.sph
-20050910_210708_33_fsp.sph
-20050913_210933_49_fsp.sph
-20050913_211649_50_fsp.sph
-20050915_210434_65_fsp.sph
-20050916_180332_68_fsp.sph
-20050918_180733_81_fsp.sph
-20050918_210841_82_fsp.sph
-20050920_212030_93_fsp.sph
-20050921_210443_99_fsp.sph
-20050923_211304_115_fsp.sph
-20050925_180713_120_fsp.sph
-20050925_180825_121_fsp.sph
-20050926_180516_125_fsp.sph
-20050926_180555_126_fsp.sph
-20050928_000254_141_fsp.sph
-20050930_210540_161_fsp.sph
-20051002_180726_170_fsp.sph
-20051007_181850_205_fsp.sph
-20051007_191217_206_fsp.sph
