--- recipes/Fisher-Callhome-Spanish/splits/test
@@ -1,20 +1,0 @@
-20051028_180633_356_fsp.sph
-20051029_211606_365_fsp.sph
-20051030_193924_371_fsp.sph
-20051101_212731_386_fsp.sph
-20051102_134901_389_fsp.sph
-20051102_180402_391_fsp.sph
-20051102_181501_393_fsp.sph
-20051103_211105_404_fsp.sph
-20051103_233456_406_fsp.sph
-20051107_184634_438_fsp.sph
-20051109_180253_445_fsp.sph
-20051109_210353_450_fsp.sph
-20051111_181045_470_fsp.sph
-20051111_182216_472_fsp.sph
-20051112_181649_485_fsp.sph
-20051113_155059_492_fsp.sph
-20051113_210221_496_fsp.sph
-20051113_214925_498_fsp.sph
-20051114_181749_505_fsp.sph
-20051115_212123_516_fsp.sph
